pyproject.toml
[project]
name = "quill"
version = "0.1.0"
description = "A generic Python framework for building document-grounded conversational assistants"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "jinja2",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
